feat: RPC block fetcher (#15)
* feat: init base logic for rpc fetcher

* feat: rpc block fetcher, move filters to internal

* move filters to internal folder
* rpc block fetcher
* add benchmarks:

goos: linux
goarch: amd64
pkg: github.com/grassrootseconomics/cic-chain-events/pkg/fetch
cpu: AMD EPYC Processor
Benchmark_RPC
Benchmark_RPC/RPC_Block_Fetcher_Benchmark
Benchmark_RPC/RPC_Block_Fetcher_Benchmark-4                   25          46000646 ns/op          221697 B/op        844 allocs/op
Benchmark_GraphQL
Benchmark_GraphQL/GraphQL_Block_Fetcher_Benchmark
Benchmark_GraphQL/GraphQL_Block_Fetcher_Benchmark-4           56          21219962 ns/op           56686 B/op         94 allocs/op
PASS
ok      github.com/grassrootseconomics/cic-chain-events/pkg/fetch       2.920s

* inline-docs: Describe RPC fetcher

+++ pkg/fetch/fetch.go
@@ -10,7 +10,7 @@ type Fetch interface {
 // Transaction reprsents a JSON object of all important mined transaction information
 type Transaction struct {
 	Block struct {
-		Number    uint   `json:"number"`
+		Number    uint64 `json:"number"`
 		Timestamp string `json:"timestamp"`
 	} `json:"block"`
 	Hash  string `json:"hash"`
@@ -23,8 +23,8 @@ type Transaction struct {
 	} `json:"to"`
 	Value     string `json:"value"`
 	InputData string `json:"inputData"`
-	Status    uint   `json:"status"`
-	GasUsed   uint   `json:"gasUsed"`
+	Status    uint64 `json:"status"`
+	GasUsed   uint64 `json:"gasUsed"`
 }
 
 // BlockResponse represents a full fetch JSON response
